pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "python-docx>=1.2.0",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
